Porting MySensors 1.4 library to Raspberry Pi
