+++ MyGateway.cpp
